requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
boto3==1.26.165
pytest==7.4.0
PyYAML==6.0